fix submodule support
Uses `rev-parse --show-toplevel` to directly get the root of the
worktree. Previously, we used `rev-parse --git-dir` to get
the path to the ".git" dir, then we tried to work out the
path of the worktree from there. That usually worked, but
failed for submodules, where the ".git" directory lives inside
the ".git/modules" directory of the parent repo, not in the
root of the child's worktree.