Fix #86 GitDir fails when constructing with a submodule (#87)
* add failing test for submodule support.

GitDir.fromExisting fails for existing and valid git submodule
repo. This commit adds a new test that demonstrates this failure.
Some minor refactoring of existing tests was also included, to
support the new test as clearly and directly as possible. All
pre-existing tests are functionally un-changed.

* fix submodule support

Uses `rev-parse --show-toplevel` to directly get the root of the
worktree. Previously, we used `rev-parse --git-dir` to get
the path to the ".git" dir, then we tried to work out the
path of the worktree from there. That usually worked, but
failed for submodules, where the ".git" directory lives inside
the ".git/modules" directory of the parent repo, not in the
root of the child's worktree.